+++ index.js
@@ -33,6 +33,7 @@ const safeObject = require('./util/safe_object');
  * @property {false|import('./rules/embeds/rsvp_button').RsvpButtonOptions} [rsvp_button] Disable RSVP buttons, or set options for the feature.
  * @property {false|import('./rules/embeds/terminal_button').TerminalButtonOptions} [terminal_button] Disable terminal buttons, or set options for the feature.
  * @property {false|import('./rules/embeds/columns').ColumnsOptions} [columns] Disable columns, or set options for the feature.
+ * @property {false} [details] Disable details.
  * @property {false} [glob] Disable glob embeds.
  * @property {false} [dns] Disable DNS lookup embeds.
  * @property {false} [asciinema] Disable Asciinema embeds.
@@ -97,6 +98,10 @@ module.exports = (md, options) => {
 
     if (optsObj.columns !== false) {
         md.use(require('./rules/embeds/columns'), safeObject(optsObj.columns));
+    }
+
+    if (optsObj.details !== false) {
+        md.use(require('./rules/embeds/details'), safeObject(optsObj.details));
     }
 
     if (optsObj.glob !== false) {